DirectShow 필터를 컴파일한 라이브러리인 strmbase.lib를 iid만 선언된 strmiids.lib로 대체 #pragma comment(lib, "strmbase.lib") -> #pragma comment(lib, "strmiids.lib")

+++ trunk/GomtosPlayer/GomtosPlayer/GomtosPlayerCtrl.cpp
@@ -341,7 +341,7 @@ DOUBLE CGomtosPlayerCtrl::GetVolume(void)
 	if (vol < -6000.)
 		vol = 0;
 	else
-		vol = (vol + 6000.) / 60.;
+		vol = (LONG) ((vol + 6000.0) / 60.);
 
 	return (DOUBLE)vol;
 }
